add templates for running other packages under Travis-CI
We want other packages (e.g., drp_stella, obs_pfs) to run the
pipeline under Travis, which involves checking this package out.
The travis.sh.TEMPLATE and travis.yml.TEMPLATE provide the
necessary machinery to do that. Unfortunately, they need to be
copied into the other packages (as travis.sh and .travis.yml).
The .TEMPLATE files here in this package will serve as the
authoritative version.