update how we build
LSST now longer produces Conda binaries (they were a pain) and we want
to use a more modern version of the LSST stack. Convert to using the
'newinstall.sh' method, which downloads binaries (if available) using
eups.

Moved the PFS build part into a separate script that we can run once
we have an LSST install.