bin: extract calib construction to separate script
Calib construction is something we want to do frequently, so make it
easy by providing a script that will do it. All the user has to do is
provide the identifiers for the bias, dark, flat, fiberTrace and arc.
Based this script on the integration test command-lines, and modified
the integration test to use the new script.