pfs_build_calibs: make all bias,dark,flat optional
Arcs have been optional for a while, and now fiberTraces are optional
because we iterate over a list which may be empty. We may as well make
all types optional, and then this script becomes more useful because we
can run it with only the things we want to create, rather than everything
(very helpful when calib construction broke halfway through).