Pass in the builder's uid/gid for the pfs account in the docker image
The pfs account in the image needs r/w access to the host filesystem,
which it will not get on anything other than the simplest
insttallation. The uid/gid in the image cannot readily be changed
dynamically, but can be set at build time. At the expense of requiring
a per-system build of the base LSST image, we can make it possible to
set the uid/gid to a host account which can r/w to host filesystems.