Fix mapping of tickets to release tag.
The mapping of tickets to release tag
using the GitHub API required an additional
query that associates the tag commit sha
with the merge commit sha.
The link between tag and tag commit,
and the merge commit to ticket,
are already in place.
With the additional query the link
between ticket and tag is complete.

An additional fix to cater for multiple tags
associated with a given commit also added.