generateCommands.py: run each calibBlock in a separate rerun-dir
Previously, a calibType of one calibBlock were constructed
in the same rerun-directory as the calibType of another calibBlock.
Without --clean option, the products of the two calibTypes were
intermingled with each other, and the latter product could not be
ingested properly. Now, the two calibTypes are constructed in separate
directories.