Revert "weekly: increase precision limit on wavelength residuals"
This reverts commit d1348f2d129351c7c11d400fcb7159d17b4aaba7.

This change was driven by an apparent decrease in wavelength
calibration precision, but it was illusory: it was caused
by retrieving the wrong detectorMap, which occurred because
the dates in the simulated data were unordered. We were
actually retrieving the SplinedDetectorMap created by the
simulator, which has a known 0.5 pixel offset (in both x and
y). By fixing the dates in the simulated data, we get the
right detectorMap and therefore can confirm a decent
wavelength calibration.