update LSST base to w_2022_17
There is a major bug in the Gen3 middleware in LSST 23, the fix
for which is not a candidate for backporting. We're going to have
to adopt a 'rolling baseline', where we update our base version
whenever we encounter a major bug. I'm a bit concerned about
getting hurt by API changes, but we have tests to catch those,
and Jim Bosch assures me that they're much better about worrying
about backwards compatibility than they used to to.

w_2022_17 is the most recent weekly snapshot that contains not
only the fix for the major bug, but also additional fixes related
to PFS (DM-34101, DM-34175, DM-34215, DM-34497).