integration_test: use new makePfsDefects command
Gen3 insists that the curated calibs are in the drp_pfs_data
package in a particular place, so we can't generate the defects
into a staging directory. Instead, modified the ingestPfsDefects
script to generate the defects in the right place, and renamed
it to makePfsDefects. It doesn't do the ingest any more (because
that's done differently for Gen2 and Gen3), so we call that
ourselves.