generateCommands: Use reduceProfiles instead of constructFiberProfiles
reduceProfiles.py is an alternative fiber profile construction code
that fits profiles to multiple images simultaneously, and is preferred
for use on Subaru data because some cobras cannot be hidden.

The syntax of generateCommand.py's input YAML files also changes with
this commit because the command line syntax of `reduceProfiles.py` is
different from that of constructFiberProfiles.