weekly: update visits for new dataset
When converting to using reduceProfiles, we discovered that
we didn't have quartzes taken with m in order to normalise
the profiles. We made a new weekly dataset (weekly-20230925)
with the m quartzes and need to update the identification of
each part of the dataset.

We don't need to update the scienceSims, since this was only
a problem with the calib construction.