weekly: don't blind find during bootstrap
The simulator exposures have so little shift w.r.t. the
detectorMap that we can simply use that as a starting point.
This saves a bunch of playing around with threshold levels
to get the right number of fibers. This is only necessary
because we don't have good simulated exposures to use for
bootstrapping, only full fiber density.